Re-factor
1. Using DRY principe to abstract out common code
2. Moved out all the database querying code into its own separate files for better management
