+++ sql_scripts/08_industry_mappings.sql
@@ -22,4 +22,7 @@ VALUES (1, 2, 1),
        (2, 2, 10),
        (2, 1, 1),
        (2, 1, 2),
-       (2, 1, 99); -- Other
+       (1, 1, 99),
+       (1, 2, 99),
+       (2, 1, 99),
+       (2, 2, 99); -- Other
